chore(lint): enable numeric-cast clippy family
Turn on cast_possible_truncation, cast_sign_loss, cast_possible_wrap, and
cast_lossless across the workspace and clear every site. A DMX/color app is
full of u8 channel values and usize buffer indices crossing into the u16/u32
DMX, wire, and DynamoDB types, where a silent truncation is a real defect.

Each flagged cast becomes an explicit decision, never an #[allow]:
- infallible widenings use From/into (Art-Net port address, node backoff
  timers, DynamoDB universe writes, u8->u16 rejection sampling)
- bounded values use checked try_from, dropping or defaulting a malformed
  record (DynamoDB universe reads) or saturating where overflow is
  unreachable (fixture span, setup counts, epoch millis)
- casts proven safe locally use expect (UNIVERSE_SIZE, PACKET_LEN, and a
  fixture end already range-checked one line above)

No wire shape or binding changes: setup compile still emits the same
ConfigFixture/ConfigChannel values, and bindings.ts is untouched.

cast_precision_loss is intentionally left off for now: its only sites are
i64 epoch-millis widened to the f64 the TS bindings require, so clearing it
means changing the timestamp representation, not the cast.

+++ services/sync-api/src/shares.rs
@@ -62,12 +62,12 @@ const CODE_LEN: usize = 10;
 /// 10%, costing roughly a bit — immaterial against a single-use 48-hour code,
 /// but the property is free to keep.)
 fn mint_code() -> Result<String, String> {
-    let span = CODE_ALPHABET.len() as u16;
+    let span = CODE_ALPHABET.len();
     let ceiling = (256 / span) * span; // largest whole multiple of the alphabet
     let mut body = String::with_capacity(CODE_LEN);
     while body.len() < CODE_LEN {
         for b in random::<CODE_LEN>()? {
-            if (b as u16) < ceiling {
+            if (b as usize) < ceiling {
                 body.push(CODE_ALPHABET[b as usize % CODE_ALPHABET.len()] as char);
                 if body.len() == CODE_LEN {
                     break;
